+++ mortar-sample/src/main/java/com/example/mortar/DemoActivity.java
@@ -125,7 +125,8 @@ public class DemoActivity extends SherlockActivity implements ActionBarOwner.Vie
 
     // activityScope may be null in case isWrongInstance() returned true in onCreate()
     if (isFinishing() && activityScope != null) {
-      activityScope.destroy();
+      MortarScope parentScope = Mortar.getScope(getApplication());
+      parentScope.destroyChild(activityScope);
       activityScope = null;
     }
   }