+++ mortar/src/main/java/mortar/PopupPresenter.java
@@ -111,9 +111,9 @@ public abstract class PopupPresenter<D extends Parcelable, R> extends Presenter<
     }
   }
 
-  @Override public void onScopeDestroyed(MortarScope scope) {
+  @Override public void onExitScope(MortarScope scope) {
     Popup<D, R> popUp = getView();
     if (popUp != null && popUp.isShowing()) popUp.dismiss(!WITH_FLOURISH);
-    super.onScopeDestroyed(scope);
+    super.onExitScope(scope);
   }
 }